+++ railties/lib/rails/application/configuration.rb
@@ -23,7 +23,7 @@ module Rails
                     :content_security_policy_nonce_generator, :content_security_policy_nonce_directives,
                     :require_master_key, :credentials, :disable_sandbox, :sandbox_by_default,
                     :add_autoload_paths_to_load_path, :rake_eager_load, :server_timing, :log_file_size,
-                    :dom_testing_default_html_version
+                    :dom_testing_default_html_version, :yjit
 
       attr_reader :encoding, :api_only, :loaded_config_version, :log_level
 
@@ -81,6 +81,7 @@ module Rails
         @rake_eager_load                         = false
         @server_timing                           = false
         @dom_testing_default_html_version        = :html4
+        @yjit                                    = false
       end
 
       # Loads default configuration values for a target version. This includes
@@ -318,6 +319,8 @@ module Rails
         when "7.2"
           load_defaults "7.1"
 
+          self.yjit = true
+
           if respond_to?(:active_job)
             active_job.enqueue_after_transaction_commit = :default
           end
@@ -329,7 +332,6 @@ module Rails
           if respond_to?(:active_record)
             active_record.postgresql_adapter_decode_dates = true
             active_record.validate_migration_timestamps = true
-            active_record.automatically_invert_plural_associations = true
           end
         when "8.0"
           load_defaults "7.2"