+++ railties/test/generators/plugin_generator_test.rb
@@ -8,11 +8,15 @@ require "rails/engine/updater"
 
 DEFAULT_PLUGIN_FILES = %w(
   .git
+  .github/dependabot.yml
+  .github/workflows/ci.yml
   .gitignore
+  .rubocop.yml
   Gemfile
   MIT-LICENSE
   README.md
   Rakefile
+  bin/rubocop
   bin/test
   bukkits.gemspec
   lib/bukkits.rb
@@ -50,7 +54,6 @@ DEFAULT_PLUGIN_FILES = %w(
   test/dummy/config/environments/production.rb
   test/dummy/config/environments/test.rb
   test/dummy/config/initializers/content_security_policy.rb
-  test/dummy/config/initializers/enable_yjit.rb
   test/dummy/config/initializers/filter_parameter_logging.rb
   test/dummy/config/initializers/inflections.rb
   test/dummy/config/initializers/permissions_policy.rb