+++ activerecord/test/cases/readonly_test.rb
@@ -79,9 +79,11 @@ class ReadOnlyTest < ActiveRecord::TestCase
 
   def test_find_with_readonly_option
     Developer.all.each { |d| assert_not d.readonly? }
+    Developer.all.tap { |rel| assert_not rel.readonly? }
     Developer.readonly(false).each { |d| assert_not d.readonly? }
     Developer.readonly(true).each { |d| assert_predicate d, :readonly? }
     Developer.readonly.each { |d| assert_predicate d, :readonly? }
+    Developer.readonly.tap { |rel| assert_predicate rel, :readonly? }
   end
 
   def test_find_with_joins_option_does_not_imply_readonly