+++ actioncable/lib/action_cable/gem_version.rb
@@ -12,7 +12,7 @@ module ActionCable
     MAJOR = 7
     MINOR = 2
     TINY  = 0
-    PRE   = "alpha"
+    PRE   = "beta1"
 
     STRING = [MAJOR, MINOR, TINY, PRE].compact.join(".")
   end